Update for QCOMPARE sensitivity to numeric types

+++ tests/src/core/testqgspointcloudattribute.cpp
@@ -77,7 +77,7 @@ void TestQgsPointCloudAttribute::testAttribute()
   QgsPointCloudAttribute attribute( QStringLiteral( "name" ), QgsPointCloudAttribute::DataType::Float );
   QCOMPARE( attribute.name(), QStringLiteral( "name" ) );
   QCOMPARE( attribute.type(), QgsPointCloudAttribute::DataType::Float );
-  QCOMPARE( attribute.size(), 4 );
+  QCOMPARE( attribute.size(), static_cast< std::size_t >( 4 ) );
 }
 
 void TestQgsPointCloudAttribute::testCollection()
@@ -85,14 +85,14 @@ void TestQgsPointCloudAttribute::testCollection()
   // test collections
   QgsPointCloudAttributeCollection collection;
   QVERIFY( collection.attributes().empty() );
-  QCOMPARE( collection.pointRecordSize(), 0 );
+  QCOMPARE( collection.pointRecordSize(), static_cast< std::size_t >( 0 ) );
   int offset = 0;
   QVERIFY( !collection.find( QStringLiteral( "test" ), offset ) );
 
   collection.push_back( QgsPointCloudAttribute( QStringLiteral( "at1" ), QgsPointCloudAttribute::DataType::Float ) );
   QCOMPARE( collection.attributes().size(), 1 );
   QCOMPARE( collection.attributes().at( 0 ).name(), QStringLiteral( "at1" ) );
-  QCOMPARE( collection.pointRecordSize(), 4 );
+  QCOMPARE( collection.pointRecordSize(), static_cast< std::size_t >( 4 ) );
   QVERIFY( !collection.find( QStringLiteral( "test" ), offset ) );
   QCOMPARE( collection.find( QStringLiteral( "at1" ), offset )->name(), QStringLiteral( "at1" ) );
   QCOMPARE( offset, 0 );
@@ -101,7 +101,7 @@ void TestQgsPointCloudAttribute::testCollection()
   QCOMPARE( collection.attributes().size(), 2 );
   QCOMPARE( collection.attributes().at( 0 ).name(), QStringLiteral( "at1" ) );
   QCOMPARE( collection.attributes().at( 1 ).name(), QStringLiteral( "at2" ) );
-  QCOMPARE( collection.pointRecordSize(), 6 );
+  QCOMPARE( collection.pointRecordSize(), static_cast< std::size_t >( 6 ) );
   QVERIFY( !collection.find( QStringLiteral( "test" ), offset ) );
   QCOMPARE( collection.find( QStringLiteral( "at1" ), offset )->name(), QStringLiteral( "at1" ) );
   QCOMPARE( offset, 0 );
@@ -113,7 +113,7 @@ void TestQgsPointCloudAttribute::testCollection()
   QCOMPARE( collection.attributes().at( 0 ).name(), QStringLiteral( "at1" ) );
   QCOMPARE( collection.attributes().at( 1 ).name(), QStringLiteral( "at2" ) );
   QCOMPARE( collection.attributes().at( 2 ).name(), QStringLiteral( "at3" ) );
-  QCOMPARE( collection.pointRecordSize(), 14 );
+  QCOMPARE( collection.pointRecordSize(), static_cast< std::size_t >( 14 ) );
   QVERIFY( !collection.find( QStringLiteral( "test" ), offset ) );
   QCOMPARE( collection.find( QStringLiteral( "at1" ), offset )->name(), QStringLiteral( "at1" ) );
   QCOMPARE( offset, 0 );
@@ -131,7 +131,7 @@ void TestQgsPointCloudAttribute::testCollection()
   QCOMPARE( collection2.attributes().at( 0 ).name(), QStringLiteral( "at1" ) );
   QCOMPARE( collection2.attributes().at( 1 ).name(), QStringLiteral( "at2" ) );
   QCOMPARE( collection2.attributes().at( 2 ).name(), QStringLiteral( "at3" ) );
-  QCOMPARE( collection2.pointRecordSize(), 14 );
+  QCOMPARE( collection2.pointRecordSize(), static_cast< std::size_t >( 14 ) );
   QVERIFY( !collection2.find( QStringLiteral( "test" ), offset ) );
   QCOMPARE( collection2.find( QStringLiteral( "at1" ), offset )->name(), QStringLiteral( "at1" ) );
   QCOMPARE( offset, 0 );
